Refactoring utils
