+++ lib/story_view.dart
@@ -1,5 +1,5 @@
+export 'controller/story_controller.dart';
+export 'utils/utils.dart';
 export 'widgets/story_image.dart';
 export 'widgets/story_video.dart';
 export 'widgets/story_view.dart';
-export 'controller/story_controller.dart';
-export 'utils.dart';
